requirements.txt
beautifulsoup4
matplotlib>=3.10.0
numpy
pillow
qrcode
reportlab
requests
streamlit